Prints summary of "RoBSA" corresponding to the input

check_setup prints summary of "RoBSA" ensemble corresponding to the specified formula, data, and priors. This function is useful for checking the ensemble configuration prior to fitting all models.

check_setup(
  formula,
  data,
  priors = NULL,
  test_predictors = NULL,
  distributions = c("exp-aft", "weibull-aft", "lnorm-aft", "llogis-aft", "gamma-aft"),
  distributions_weights = rep(1, length(distributions)),
  prior_beta_null = get_default_prior_beta_null(),
  prior_beta_alt = get_default_prior_beta_alt(),
  prior_factor_null = get_default_prior_factor_null(),
  prior_factor_alt = get_default_prior_factor_alt(),
  prior_intercept = get_default_prior_intercept(),
  prior_aux = get_default_prior_aux(),
  chains = 3,
  sample = 5000,
  burnin = 2000,
  adapt = 500,
  thin = 1,
  parallel = FALSE,
  autofit = TRUE,
  autofit_control = set_autofit_control(),
  convergence_checks = set_convergence_checks(),
  save = "all",
  seed = NULL,
  silent = FALSE,
  rescale_data = FALSE,
  models = FALSE,
  ...
)

Arguments

formula

formula for the survival model

data

data frame containing the data

priors

names list of prior distributions for each predictor. It allows users to specify both the null and alternative hypothesis prior distributions by assigning a named list (with "null" and "alt" object) to the predictor

test_predictors

vector of predictor names to be tested with Bayesian model-averaged testing. Defaults to NULL, no parameters are tested.

distributions

distributions of parametric survival models

distributions_weights

prior odds for the competing distributions

prior_beta_null

default prior distribution for the null hypotheses of continuous predictors

prior_beta_alt

default prior distribution for the alternative hypotheses of continuous predictors

prior_factor_null

default prior distribution for the null hypotheses of categorical predictors

prior_factor_alt

default prior distribution for the alternative hypotheses of categorical predictors

prior_intercept

named list containing prior distribution for the intercepts (with names corresponding to the distributions)

prior_aux

named list containing prior distribution for the auxiliary parameters (with names corresponding to the distributions)

chains

a number of chains of the MCMC algorithm.

sample

a number of sampling iterations of the MCMC algorithm. Defaults to 5000.

burnin

a number of burnin iterations of the MCMC algorithm. Defaults to 2000.

adapt

a number of adaptation iterations of the MCMC algorithm. Defaults to 500.

thin

a thinning of the chains of the MCMC algorithm. Defaults to 1.

parallel

whether the individual models should be fitted in parallel. Defaults to FALSE. The implementation is not completely stable and might cause a connection error.

autofit

whether the model should be fitted until the convergence criteria (specified in autofit_control) are satisfied. Defaults to TRUE.

autofit_control

allows to pass autofit control settings with the set_autofit_control() function. See ?set_autofit_control for options and default settings.

convergence_checks

automatic convergence checks to assess the fitted models, passed with set_convergence_checks() function. See ?set_convergence_checks for options and default settings.

save

whether all models posterior distributions should be kept after obtaining a model-averaged result. Defaults to "all" which does not remove anything. Set to "min" to significantly reduce the size of final object, however, some model diagnostics and further manipulation with the object will not be possible.

seed

a seed to be set before model fitting, marginal likelihood computation, and posterior mixing for reproducibility of results. Defaults to NULL - no seed is set.

silent

do not print the results.

rescale_data

whether continuous predictors should be rescaled prior to estimating the model. Defaults to FALSE.

models

should the models' details be printed.

...

additional arguments.

Value

check_setup invisibly returns list of summary tables.

See also